# [XamlC] use XmlReader for detecting XAML and x:Class