--- a/Xamarin.Forms.Xaml.UnitTests/Validation/MissingXClass.xaml
+++ b/Xamarin.Forms.Xaml.UnitTests/Validation/MissingXClass.xaml
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<ContentPage 
+	xmlns="http://xamarin.com/schemas/2014/forms"
+	xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml">
+</ContentPage>