--- a/Xamarin.Forms.Xaml.UnitTests/Validation/NotXaml.xaml
+++ b/Xamarin.Forms.Xaml.UnitTests/Validation/NotXaml.xaml
@@ -1,0 +1,3 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<foo>
+</foo>