--- a/Xamarin.Forms.Xaml.UnitTests/X2006Namespace.xaml
+++ b/Xamarin.Forms.Xaml.UnitTests/X2006Namespace.xaml
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             x:Class="Xamarin.Forms.Xaml.UnitTests.X2006Namespace">
+</ContentPage>