@StephaneDelcroix 2020-03-11
I'm surprised that XamlG doesn't show up in your performances then, as it also uses XmlDocument and then does some XPath queries too...