@jonathanpeppers 2020-03-11
> I'm surprised that XamlG doesn't show up in your performances

Yeah, I'm using this "hot path" feature:

![image](https://user-images.githubusercontent.com/840039/76422267-e9e8ef80-6372-11ea-92b5-df13ff0394ad.png)

Not sure why XamlG isn't showing up, this is how long it actually took:

![image](https://user-images.githubusercontent.com/840039/76422294-f8370b80-6372-11ea-9e4c-67f40bcd4a5d.png)

37ms is pretty quick, but I'll take a look and see if it has something similar.